@@ -46,7 +46,7 @@ var (
 		Name:          "Home",
 		PresetID:      1,
 		Image:         "ccc47fae-847c-4a91-8a65-b26cbae6fbe2.jpg",
-		LectureHallId: LectureHall.ID,
+		LectureHallID: LectureHall.ID,
 		IsDefault:     false,
 	}
 	CourseFPV = model.Course{
